Extract export function into separate file
